+++ CHANGELOG.md
@@ -17,9 +17,11 @@ All user visible changes to this project will be documented in this file. This p
 ### Changed
 
 - Upgraded [OpenAL] library to [1.25.0][openal-1.25.0] version. ([#275])
+- Upgraded [libwebrtc] to [143.0.7499.192] version. ([todo])
 
 [#275]: https://github.com/instrumentisto/medea-flutter-webrtc/pull/275
 [#277]: https://github.com/instrumentisto/medea-flutter-webrtc/pull/277
+[143.0.7499.192]: https://github.com/instrumentisto/libwebrtc-bin/releases/tag/143.0.7499.192
 [openal-1.25.0]: https://github.com/kcat/openal-soft/releases/tag/1.25.0